make StreamCorruption_IsDetected run for gzip only
other types of stream can't detect corruption properly
