+++ src/libraries/System.IO.Compression/tests/CompressionStreamUnitTests.Gzip.cs
@@ -5,8 +5,8 @@ using System.Buffers;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using System.IO.Compression.Tests;
 using Xunit;
-using Xunit.Sdk;
 
 namespace System.IO.Compression
 {
@@ -271,6 +271,50 @@ namespace System.IO.Compression
             {
                 compressor.WriteAsync(new ReadOnlyMemory<byte>(new byte[1])).AsTask().Wait();
                 Assert.True(compressor.WriteArrayInvoked);
+            }
+        }
+
+        [Fact]
+        public void StreamCorruption_IsDetected()
+        {
+            var source = Enumerable.Range(0, 64).Select(i => (byte)i).ToArray();
+            var buffer = new byte[64];
+            byte[] compressedData;
+            using (var compressed = new MemoryStream())
+            using (Stream compressor = CreateStream(compressed, CompressionMode.Compress))
+            {
+                foreach (byte b in source)
+                {
+                    compressor.WriteByte(b);
+                }
+
+                compressor.Dispose();
+                compressedData = compressed.ToArray();
+            }
+
+            for (int byteToCorrupt = 0; byteToCorrupt < compressedData.Length; byteToCorrupt++)
+            {
+                // corrupt the data
+                compressedData[byteToCorrupt]++;
+
+                using (var decompressedStream = new MemoryStream(compressedData))
+                {
+                    using (Stream decompressor = CreateStream(decompressedStream, CompressionMode.Decompress))
+                    {
+                        try
+                        {
+                            while (ZipFileTestBase.ReadAllBytes(decompressor, buffer, 0, buffer.Length) != 0) { };
+
+                            Assert.Equal(source, buffer);
+                        }
+                        catch (InvalidDataException)
+                        {
+                        }
+                    }
+                }
+
+                // restore the data
+                compressedData[byteToCorrupt]--;
             }
         }
